refactor user forms
