--- src/resources/users/UserCreate.tsx
@@ -1,10 +1,0 @@
-import { Create } from 'react-admin';
-import UserForm from '../../components/UserFrom';
-
-export default function UserCreate() {
-	return (
-		<Create>
-			<UserForm />
-		</Create>
-	);
-}
