--- src/resources/users/UserEdit.tsx
@@ -1,10 +1,0 @@
-import { Edit } from 'react-admin';
-import UserForm from '../../components/UserFrom';
-
-export default function UserEdit() {
-	return (
-		<Edit>
-			<UserForm />
-		</Edit>
-	);
-}
